unsigned data -> signed

+++ HW1/miniknn.py
@@ -22,7 +22,7 @@ def kNNClassify(newInput, dataSet, labels, k):
     ########################
      # Calculate the L2 distance
     def getL2Distance(pair1, pair2):
-        return np.sqrt(np.sum(np.power(np.subtract(pair1,pair2),2)))
+        return np.sqrt(np.sum(np.power(np.subtract(pair1,pair2, dtype=int),2)))
 
     # Get the most common element in a given list
     def getMostCommon(labels):
